Adds clean-up logic.
Adds logic around file remove for easier flipping the switch

+++ migration/rnaseq-wf_aln_stats.py
@@ -8,6 +8,7 @@ sys.path.insert(0, "../src")
 from ncbi_remap.parser import parse_bamtools_stats, parse_samtools_stats
 
 
+CLEAN_UP = False
 RNASEQ_PATH = Path("../output/rnaseq-wf/samples")
 
 Files = namedtuple("Files", "srx idx idxstats samtools bamtools output")
@@ -33,11 +34,13 @@ def main():
             else:
                 print(files.samtools, files.bamtools)
 
-            files.samtools.unlink()
-            files.bamtools.unlink()
+            if CLEAN_UP:
+                files.samtools.unlink()
+                files.bamtools.unlink()
 
         if files.idxstats.exists():
-            files.idxstats.unlink()
+            if CLEAN_UP:
+                files.idxstats.unlink()
 
 
 def convert_table(files):
