+++ migration/rnaseq-wf_hisat2.py
@@ -9,6 +9,7 @@ import pandas as pd
 sys.path.insert(0, "../src")
 from ncbi_remap.parser import parse_hisat2
 
+CLEAN_UP = False
 RNASEQ_PATH = Path("../output/rnaseq-wf/samples")
 
 Files = namedtuple("Files", "srr idx log bam bai output")
@@ -35,7 +36,8 @@ def main():
         )
         if files.log.exists():
             convert_log(files)
-            Path(files.log).unlink()
+            if CLEAN_UP:
+                Path(files.log).unlink()
 
     remove_alignment_bad()
 
@@ -68,10 +70,10 @@ def convert_log(files):
         alignment_bad_path.mkdir(exist_ok=True)
         Path(alignment_bad_path, files.srr).touch()
 
-        if Path(files.bam).exists():
+        if Path(files.bam).exists() and CLEAN_UP:
             Path(files.bam).unlink()
 
-        if Path(files.bai).exists():
+        if Path(files.bai).exists() and CLEAN_UP:
             Path(files.bai).unlink()
 
 
@@ -80,7 +82,7 @@ def remove_alignment_bad():
     for pth in Path("../output/rnaseq-wf/samples/").glob("**/SRR*"):
         if pth.name in srrs:
             for file_pth in pth.iterdir():
-                if not "fastq_screen" in file_pth.name:
+                if not "fastq_screen" in file_pth.name and CLEAN_UP:
                     file_pth.unlink()
